VideoPress: Extract VideoDetails from VideoCard
