+++ projects/packages/videopress/src/client/admin/components/video-card/stories/VideoCard.mdx
@@ -22,7 +22,3 @@ It is composed of two simpler components: VideoThumbnail and VideoDetails compon
 # VideoThumbnail
 
 <Story id="packages-videopress-video-card--video-thumbnail" />
-
-# VideoDetails
-
-<Story id="packages-videopress-video-card--video-details" />
