+++ projects/packages/videopress/src/client/admin/components/video-card/style.module.scss
@@ -71,14 +71,6 @@
 	}
 }
 
-.details {
-	color: var( --jp-black-80 );
-	padding: calc( var( --spacing-base ) * 3 ); // 24px;
-	display: flex;
-	flex-direction: column;
-	gap: calc( var( --spacing-base ) * 2 );;
-}
-
 .video-card__wrapper {
 	--videopress-video-item-max-width: 392px; // VideosGrid: two columns.
 	max-width: var( --videopress-video-item-max-width );
