+++ projects/packages/videopress/src/client/admin/components/video-details/style.module.scss
@@ -1,0 +1,7 @@
+.details {
+	color: var( --jp-black-80 );
+	padding: calc( var( --spacing-base ) * 3 ); // 24px;
+	display: flex;
+	flex-direction: column;
+	gap: calc( var( --spacing-base ) * 2 );;
+}
