+++ projects/packages/videopress/src/client/admin/components/video-details/types.ts
@@ -1,0 +1,16 @@
+export type VideoDetailsProps = {
+	/**
+	 * Video filename.
+	 */
+	filename: string;
+
+	/**
+	 * Video source file URL.
+	 */
+	src: string;
+
+	/**
+	 * Video uploaded date
+	 */
+	uploadDate: string;
+};
